Improve interpolation when the client rate limit is less than the server's tick rate. Fix some states assuming their packets will always be sent, regardless of the rate limit.

+++ include/nrg_client_state.h
@@ -66,7 +66,7 @@ public:
 	
 	void sendMessage(const MessageBase& m);
 	
-	float getInterpTimer() const;
+	double getInterpTimer() const;
 	const ClientStats& getClientStats() const;
 
 	void startRecordingReplay(const char* filename);
